Inline the size and data pointers in QByteArray

+++ src/corelib/tools/qstring.h
@@ -224,8 +224,8 @@ public:
     { qSwap(d, other.d); return *this; }
 #endif
     inline void swap(QString &other) Q_DECL_NOTHROW { qSwap(d, other.d); }
-    inline int size() const { return d.size; }
-    inline int count() const { return d.size; }
+    inline int size() const { return int(d.size); }
+    inline int count() const { return int(d.size); }
     inline int length() const;
     inline bool isEmpty() const;
     void resize(int size);
@@ -870,7 +870,7 @@ public:
 inline QString::QString(QLatin1String aLatin1) : d(fromLatin1_helper(aLatin1.latin1(), aLatin1.size()))
 { }
 inline int QString::length() const
-{ return d.size; }
+{ return int(d.size); }
 inline const QChar QString::at(int i) const
 { Q_ASSERT(uint(i) < uint(size())); return d.b[i]; }
 inline const QChar QString::operator[](int i) const
@@ -1076,7 +1076,7 @@ inline QString::~QString() { if (!d.d->deref()) Data::deallocate(d.d); }
 inline void QString::reserve(int asize)
 {
     if (d.d->needsDetach() || asize >= capacity())
-        reallocData(qMax(asize, size()) + 1u);
+        reallocData(uint(qMax(asize, size())) + 1u);
 
     // we're not shared anymore, for sure
     d.d->flags |= Data::CapacityReserved;
@@ -1090,7 +1090,7 @@ inline void QString::squeeze()
         reallocData(uint(d.size) + 1u);
 
     // we're not shared anymore, for sure
-    d.d->flags &= ~Data::CapacityReserved;
+    d.d->flags &= uint(~Data::CapacityReserved);
 }
 
 inline QString &QString::setUtf16(const ushort *autf16, int asize)
